pyproject.toml
[project]
authors = [{ name = "Alejandro Ouslan", email = "alejandro.ouslan@upr.edu" }]
maintainers = [
  { name = "Yamilette Alemany", email = "yamilette.alemany@upr.edu" },
  { name = "Adriana I. Carrasquillo Cruz", email = "adriana.carrasquillo1@upr.edu" },
  { name = "Fernando J. Sierra Aponte", email = "fernando.sierra3@upr.edu" },
  { name = "Bryan D. Vega Gonzalez", email = "bryan.vega7@upr.edu" },
  { name = "Edzel Enrique Ortiz Padilla", email = "edzel.ortiz@upr.edu" },
]
license = { text = "GPL-3.0-or-later" }
name = "jp-webapp"
version = "1.4.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "django>=5.1.5",
  "numpy>=2.1.3",
  "pandas>=2.2.3",
  "plotly>=5.24.1",
  "polars>=1.16.0",
  "psycopg2-binary>=2.9.10",
  "pyarrow>=18.1.0",
  "pytest>=8.3.4",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "scipy>=1.14.1",
  "sqlalchemy>=2.0.36",
  "sqlmodel>=0.0.22",
  "tk>=0.1.0",
  "tqdm>=4.67.1",
  "whitenoise>=6.8.2",
]
